requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
PyMySQL==1.1.0
openai==1.3.5
requests==2.31.0
Flask-JWT-Extended==4.5.3
passlib==1.7.4
email-validator==2.1.0.post1
PyJWT==2.8.0
